Task 1 SchedulingSystem accessor methods complete
Task 1 has been complete. The accessor methods getSystemTime(), getNumProcesses(), isCpuIdle(), and getRunningProcessName() all have the correct prototypes, Doxygen documentation, and implementation completed allowing all subtasks of Task 1 to be passed.

The methods getSystemTime() and getNumProcesses() simply return the integer values of the current systemTime and numProcesses variables. The isCpuIdle() method checks if cpu is IDLE, if it is then true is returned, otherwise false is returned. The getRunningProcessName() method first checks if the cpu is IDLE with the isCpuIdle() method returning a string "IDLE" if true, otherwise the method will get the name of the current process running on the cpu.