Task 3 dispatchCpuIfIdle() is complete
Task 3 is completed. The method dispatchCpuIfIdle() has the correct Dyoxegen documentation, prototype definition, and implementation to pass all Task 3 tests.

The method checks if the CPU is idle through the isCpuIdle() method, if it is then the method calls upon the policy object to dispatch the next process. When the policy object returns pid of the next dispatched process it is set as the running process in the CPU. After setting it as the running process the method checks if the process has been started before, if it hasn't then the startTime is set to the current system time through the getSystemTime() method.