Task 4 checkProcessFinished() is complete
Task 4 is completed. The method checkProcessFinished() has its Dyoxygen documentation, prototype definition, and implementaion completed correctly with all Task 4 test passing successfully.

The method first checks if the CPU is idle, if it isn't then the current process in the CPU is checked if its usedTime is greater than or equal to its serviceTime to determine if it has finished. If the process is finished then the method marks the process's endTime with the current system time and marks the process as done (true). Finally, the CPU is set to idle to allow another process to be dispatched.