Task 6_2 RR implementation completed
Task 6_2 completed. The system tests for Round Robin (RR) have all been passed with the changes made to the RR policy methods newProcess(), dispatch(), preempt(), and resetPolicy() all having the proper Dyoxygen documentation.

The changes made within the RRSchedulingPolicy.hpp files follow the creation of new private variables including 'quantum" which keeps track of the simulation's set time slice quantum. The "quantumClock" variable tracks how long a process's quantum lasts throughout the CPU cycles. Finally, the "runningPid" acted as a temp variable holding the currently running process on the CPU, allowing the process to be added back to the ready queue if it was preempted due to a quantum timer.

The changes made within the RRSchedulingPolicy.cpp files follow changes to allow a parameter to be passed into the RRSchedulingPolicy() method, this allows us to get the time quantum of the particular simulation. The dispatch() method only dispatches if there any processes in the ready queue, if a process is dispatched it placed in the runningPid variable, popped off the ready queue, sets the quantumClock for the process, simulates one CPU cycle by decrementing the quantumClock, and finally returns the Pid of the dispatched process. In the preempt() method if the quantumClock is not zero it is decremented to simulate a CPU cycle, otherwise if the quantumClock reaches zero the process is pushed back onto the ready queue by using the runningPid variable and then returns true to say it should be preempted. The last method resetPolicy() simply sets the ready queue to be empty and sets the quantumClock to the quantum time.